Modifica del RF2 ed implementazione dello stesso
