+++ ALEC_Chatbot/src/chatbot/script/debug.py
@@ -1,0 +1,8 @@
+# Flag di debug centralizzati
+DEBUG_JUDGE = True
+DEBUG_MODE  = True
+DEBUG_CTX   = False
+DEBUG_SAVE  = False
+
+def dbg(msg: str):
+    print(msg)
